requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-SocketIO
mysql-connector-python
python-dotenv
werkzeug
Pillow
google-generativeai
google-cloud-aiplatform
twilio